fixed warnings
git-svn-id: https://svn.opendtect.org/od/trunk@27491 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Algo/smoother2d.h
@@ -219,7 +219,7 @@ bool Smoother2D<T>::execute()
 	}
 
 	if ( weightsum>1 )
-	    mPointerOperation( float, window_->getData(), /= weightsum,
+	    mPointerOperation( float, window_->getData(), /= (float)weightsum,
 		        window_->info().getTotalSz(), ++ );
 
 	convolver_.setY( *window_, false );
